+++ compose.yml
@@ -42,6 +42,7 @@ services:
       PORT: 8080
       Mongo__DatabaseUri: mongodb://mongodb:27017/?directConnection=true
       SQS_ENDPOINT: http://localstack:4566
+      AWS_EMF_ENVIRONMENT: Local
     ports:
       - "8080:8080"
     healthcheck:
