+++ src/Comparer/appsettings.Development.json
@@ -1,5 +1,4 @@
 {
-  "AWS_EMF_ENABLED": false,
   "Mongo": {
     "DatabaseUri": "mongodb://127.0.0.1:27017/?directConnection=true",
     "DatabaseName": "trade-imports-decision-comparer"
