+++ src/Comparer/appsettings.IntegrationTests.json
@@ -1,4 +1,5 @@
 {
+  "AWS_EMF_ENABLED": false,
   "Mongo": {
     "DatabaseUri": "mongodb://127.0.0.1:27017",
     "DatabaseName": "trade-imports-decision-comparer-integration-tests"
