+++ src/Comparer/appsettings.json
@@ -1,6 +1,4 @@
 {
-  "AWS_EMF_ENABLED": true,
-  "AWS_EMF_NAMESPACE": "DEFRA/TRADE_IMPORTS/DECISION_COMPARER",
   "Mongo": {
     "DatabaseUri": "mongodb://set-automatically-when-deployed/admin?authSource=$external&authMechanism=MONGODB-AWS",
     "DatabaseName": "trade-imports-decision-comparer"
